Report exceptions in SqliteMultithread.
This is a first pass implementation, requesting feedback.

Because of the asynchronous "fire and forget" nature of calling
SqliteMultithread.execute(), there is no way for the calling
thread to be made aware of an exception in the inner thread.

We handle such exception gracefully, logging to level ERROR
the inner exception, the exception only, *and* the outer
stacktrace (the true offending code from the caller's thread).

**This might have a slight performance impact.**

For the ``select*`` family of methods, the inner exception is raised
in by calling thread before returning, but for "fire and forget"
calls of the ``execute*`` family of methods, such as
``db['key'] = 'value'``, we may only throw the exception:

 - on commit: somewhat mirroring asyncio's exceptions thrown
   when accessing the result of a Future.
 - on close or __del__, (which now awaits confirmation
   before returning, allowing exceptions to be caught).
 - at any subsequent statement after the time
   that the inner exception occurred.

commit() is made blocking by default (awaiting confirmation),
except by implied commit when using by autocommit=True.